fix(devserver): Adjust for delayed starting
