+++ src/Uno.UI.DevServer.Cli/Program.cs
@@ -1,7 +1,10 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Console;
+using Uno.UI.DevServer.Cli.Helpers;
 using Uno.UI.DevServer.Cli.Logging;
+using Uno.UI.DevServer.Cli.Mcp;
+using System.Globalization;
 
 namespace Uno.UI.DevServer.Cli;
 
@@ -14,16 +17,26 @@ internal class Program
 			Console.WriteLine("Usage: dnx -y uno.devserver [options] [command]");
 			Console.WriteLine();
 			Console.WriteLine("Options:");
-			Console.WriteLine("  --help, -h               Show this help message and exit");
-			Console.WriteLine("  --log-level, -l <level>  Set the log level (Trace, Debug, Information, Warning, Error, Critical, None). Default is Information.");
-			Console.WriteLine("  --mcp                    Start in MCP STDIO mode");
+			Console.WriteLine(" --help, -h               Show this help message and exit");
+			Console.WriteLine(" --log-level, -l <level>  Set the log level (Trace, Debug, Information, Warning, Error, Critical, None). Default is Information.");
+			Console.WriteLine(" --file-log, -fl <path>   Enable file logging to the provided file path (supports {Date} token). Required path argument.");
+			Console.WriteLine(" --mcp                    Start in MCP STDIO mode");
 			Console.WriteLine();
 			Console.WriteLine("Commands:");
-			Console.WriteLine("  start                    Start the DevServer for the current folder");
-			Console.WriteLine("  stop                     Stop the DevServer for the current folder");
-			Console.WriteLine("  list                     List active DevServer instances");
+			Console.WriteLine(" start                    Start the DevServer for the current folder");
+			Console.WriteLine(" stop                     Stop the DevServer for the current folder");
+			Console.WriteLine(" list                     List active DevServer instances");
 			Console.WriteLine();
 			return 0;
+		}
+
+		// Parse common options up-front
+		var logLevel = GetLogLevelFromArgs(args);
+		var fileLogPath = GetFileLogPathFromArgs(args, out var fileLogPathError);
+		if (fileLogPathError is not null)
+		{
+			Console.Error.WriteLine(fileLogPathError);
+			return 1;
 		}
 
 		var services = new ServiceCollection();
@@ -35,9 +48,31 @@ internal class Program
 				options.LogToStandardErrorThreshold = LogLevel.Trace;
 			});
 			builder.AddConsoleFormatter<CleanConsoleFormatter, ConsoleFormatterOptions>();
-			builder.SetMinimumLevel(GetLogLevelFromArgs(args));
+
+			builder.SetMinimumLevel(logLevel);
+
+			if (fileLogPath is not null)
+			{
+				try
+				{
+					var dir = Path.GetDirectoryName(fileLogPath);
+					if (!string.IsNullOrEmpty(dir))
+					{
+						Directory.CreateDirectory(dir);
+						builder.AddFile(fileLogPath, minimumLevel: logLevel);
+					}
+				}
+				catch (Exception e)
+				{
+					Console.Error.WriteLine($"Failed to initialize file logging for '{fileLogPath}': {e.Message}");
+				}
+			}
 		});
 		services.AddSingleton<CliManager>();
+		services.AddSingleton<UnoToolsLocator>();
+		services.AddSingleton<DevServerMonitor>();
+		services.AddSingleton<McpProxy>();
+		services.AddSingleton<McpClientProxy>();
 
 		using var sp = services.BuildServiceProvider();
 		var manager = sp.GetRequiredService<CliManager>();
@@ -62,5 +97,29 @@ internal class Program
 #else
 		return LogLevel.Warning;
 #endif
+	}
+
+	private static string? GetFileLogPathFromArgs(string[] args, out string? error)
+	{
+		error = null;
+		for (int i = 0; i < args.Length; i++)
+		{
+			if (args[i] == "-fl")
+			{
+				if (i == args.Length - 1)
+				{
+					error = "Missing file path after -fl option.";
+					return null;
+				}
+				var path = args[i + 1];
+				if (string.IsNullOrWhiteSpace(path))
+				{
+					error = "File log path cannot be empty.";
+					return null;
+				}
+				return path;
+			}
+		}
+		return null;
 	}
 }
